drm/i915/xelpg: Extend some workarounds/tuning to gfx version 12.74
Some of our existing Xe_LPG workarounds and tuning are also applicable
to the version 12.74 variant.  Extend the condition bounds accordingly.
Also fix the comment on Wa_14018575942 while we're at it.

FIXME:  Wa_14016712196 was not implemented properly upstream.  Once
that's fixed & backported, we should just be extending the upper bound
of a range rather than adding a separate condition line.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>